Refactor echo actor and WebSocket transport for multi-protocol tests
Replace HTTPClient.NewWebSocket with three narrower accessors —
Transport(), Headers(), and BaseURL() — so each transport can build
its own connection using shared credentials without depending on a
WebSocket-specific helper.

Move the WebSocket creation logic into newWebSocketClient in the
rcprotocoltest component, and update runEchoLoop to accept a runCount
uint64 (replacing the reconnections counter) which is forwarded as the
X-Echo-Run-Count request header. Remove the RunEchoTest outer retry
loop; retry responsibility now belongs to the actor and the
RunTransportTests coordinator.

Rename webSocketTestActor to echoTestActor and wire it to
RunTransportTests. Update Stop to use sync.Once for safe repeated
calls. Thread runCount through the actor's tick loop and into each
test invocation.

Move TestNewWebSocket from api/http_test.go into the component test
file, updating it to use newWebSocketClient. Update reconnection
header assertions from X-Echo-Reconnections to X-Echo-Run-Count.
