+++ pkg/config/remote/api/http.go
@@ -15,15 +15,11 @@ import (
 	"maps"
 	"net/http"
 	"net/url"
-	"path"
 	"runtime"
-	"strings"
 	"sync"
 	"time"
 
 	"google.golang.org/protobuf/proto"
-
-	"github.com/gorilla/websocket"
 
 	"github.com/DataDog/datadog-agent/pkg/config/model"
 	pbgo "github.com/DataDog/datadog-agent/pkg/proto/pbgo/core"
@@ -267,63 +263,29 @@ func (c *HTTPClient) UpdateAPIKey(apiKey string) {
 	c.headerLock.Unlock()
 }
 
-// NewWebSocket connects to the RC WebSocket backend and returns a new WebSocket
-// connection or a connection / handshake error.
+// Transport returns a pointer to the transport (TLS + Proxy) configuration.
 //
-// The "endpointPath" specifies the resource path to connect to, which is
-// appended to the client baseURL.
-// Any headers in "extraHeaders" are added into the handshake request
-// after the standard auth headers
-func (c *HTTPClient) NewWebSocket(ctx context.Context, endpointPath string, extraHeaders http.Header) (*websocket.Conn, error) {
-	// Extract the TLS & Proxy configuration from the HTTP client.
+// Callers MUST NOT modify the returned data.
+func (c *HTTPClient) Transport() (*http.Transport, error) {
 	transport, ok := c.client.Transport.(*http.Transport)
 	if !ok {
 		return nil, errors.New("failed to extract transport config from HTTP client")
 	}
+	return transport, nil
+}
 
-	dialer := &websocket.Dialer{
-		HandshakeTimeout: 30 * time.Second,
-		TLSClientConfig:  transport.TLSClientConfig,
-		Proxy:            transport.Proxy,
-	}
-
-	// The WebSocket request MUST include the same auth credentials as the plain
-	// HTTP requests.
+// Headers returns a copy of the HTTP headers configured for RC backend
+// requests, including API credentials.
+func (c *HTTPClient) Headers() http.Header {
 	c.headerLock.RLock()
 	headers := maps.Clone(c.header)
 	c.headerLock.RUnlock()
+	return headers
+}
 
-	// Add any per-connection headers on top of the shared auth headers
-	maps.Copy(headers, extraHeaders)
-
-	// Parse the "base URL" the client uses to connect to RC.
-	url, err := url.Parse(c.baseURL)
-	if err != nil {
-		return nil, err
-	}
-	// Append the specific path to the WebSocket resource.
-	url.Path = path.Join(url.Path, endpointPath)
-	// Change the protocol to use websockets.
-	switch strings.ToLower(url.Scheme) {
-	case "http":
-		url.Scheme = "ws"
-	case "https":
-		url.Scheme = "wss"
-	}
-
-	log.Debugf("connecting to websocket endpoint %s", url.String())
-
-	// Send the HTTP request, wait for the upgrade response and then perform the
-	// WebSocket handshake.
-	conn, resp, err := dialer.DialContext(ctx, url.String(), headers)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open websocket connection: %s", err)
-	}
-	_ = resp.Body.Close()
-
-	log.Debug("websocket connected")
-
-	return conn, nil
+// BaseURL returns the HTTP endpoint address to the RC backend.
+func (c *HTTPClient) BaseURL() (*url.URL, error) {
+	return url.Parse(c.baseURL)
 }
 
 func (c *HTTPClient) addHeaders(req *http.Request) {
